requirements.txt
# Strata — Python 依赖列表
# 安装：pip install -r requirements.txt

# ── 核心框架 ──────────────────────────────────────────────────────────────────
agentscope>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# ── 向量存储 ──────────────────────────────────────────────────────────────────
chromadb>=0.4.0

# ── AD-LTM 长期记忆模块 ───────────────────────────────────────────────────────
sentence-transformers>=2.7.0       # BGE Embedding + BGE Reranker
rank-bm25>=0.2.2                   # BM25Okapi 倒排索引
langchain-text-splitters>=0.2.0    # MarkdownHeaderTextSplitter + RecursiveCharacterTextSplitter
cachetools>=5.3.0                  # LRUCache
transformers>=4.40.0               # tokenizer（用于 TER 计算）

# ── 可选：中文分词（提升 BM25 召回质量） ──────────────────────────────────────
# jieba>=0.42.1

# ── 工具链 ────────────────────────────────────────────────────────────────────
tavily-python>=0.3.0               # Tavily 搜索 API

# ── 文件解析 ──────────────────────────────────────────────────────────────────
pymupdf>=1.23.0                    # PDF 解析（优先）
pdfplumber>=0.10.0                 # PDF 解析（备用）
python-docx>=1.0.0                 # Word 文档解析
Pillow>=10.0.0                     # 图片处理

# ── 异步 ──────────────────────────────────────────────────────────────────────
aiofiles>=23.0.0

# ── 文件监听（长期记忆 FileWatcher） ─────────────────────────────────────────
watchfiles>=0.21.0

# ── 实验评估 ──────────────────────────────────────────────────────────────────
# 以上依赖已覆盖实验所需，无需额外安装
